Hook ViewPager adapter to ViewPager2
